Project finished
There is no route to take you to a profile page included in the Codecademy files.

If you want to test your logout function, you need to manually append /users/logout in the url bar.

To ensure it's working, I added a console log after the call to req.logout() to log (‘Logged out’) to the terminal.

There are a few bugs with this project as a whole as other users have logged similar complaints.